move value tokens to seperate folder
