+++ Code/ArgumentSystem/Arguments/FloatArgument.cs
@@ -3,6 +3,7 @@ using SER.Code.ArgumentSystem.BaseArguments;
 using SER.Code.Exceptions;
 using SER.Code.Helpers.ResultSystem;
 using SER.Code.TokenSystem.Tokens;
+using SER.Code.TokenSystem.Tokens.ValueTokens;
 using SER.Code.ValueSystem;
 using Random = UnityEngine.Random;
