+++ Code/ArgumentSystem/Arguments/TextArgument.cs
@@ -4,6 +4,7 @@ using SER.Code.Extensions;
 using SER.Code.Helpers.ResultSystem;
 using SER.Code.TokenSystem.Tokens;
 using SER.Code.TokenSystem.Tokens.Interfaces;
+using SER.Code.TokenSystem.Tokens.ValueTokens;
 using SER.Code.ValueSystem;
 
 namespace SER.Code.ArgumentSystem.Arguments;
