+++ Code/TokenSystem/Tokens/ValueTokens/DurationToken.cs
@@ -2,7 +2,7 @@
 using SER.Code.ScriptSystem;
 using SER.Code.ValueSystem;
 
-namespace SER.Code.TokenSystem.Tokens;
+namespace SER.Code.TokenSystem.Tokens.ValueTokens;
 
 public class DurationToken : LiteralValueToken<DurationValue>
 {
