+++ Code/TokenSystem/Tokens/ValueTokens/NumberToken.cs
@@ -1,7 +1,7 @@
 ﻿using SER.Code.ScriptSystem;
 using SER.Code.ValueSystem;
 
-namespace SER.Code.TokenSystem.Tokens;
+namespace SER.Code.TokenSystem.Tokens.ValueTokens;
 
 public class NumberToken : LiteralValueToken<NumberValue>
 {
